Add entirely new late_command value
This took a long while to get right. Each command needs in-target appended
at the front to be ran in the target or what will be the actual root
filesystem.

That said, pipes did not seem to work and requires to be invoked in a
separate shell (whose root is based on in-target).

&& were replaced because I am unsure how in-target interpretes the
entire expression. This may change.